Refactor collaborative state into modular hooks
